clap: Replace deprecated `parse_from_occurrences` with `ArgAction::Count`
Fixes these warnings when running `cargo check --features clap/deprecated`:
warning: use of deprecated function `<Opts as clap::Args>::augment_args::parse_from_occurrences`: Replaced with `#[clap(action = ArgAction::Count)]` with a field type of `u8`
   --> src/main.rs:139:43
    |
139 |     #[clap(short, long = "verbose", parse(from_occurrences))]
    |                                           ^^^^^^^^^^^^^^^^
    |
    = note: `#[warn(deprecated)]` on by default

warning: use of deprecated function `<Opts as clap::Args>::augment_args_for_update::parse_from_occurrences`: Replaced with `#[clap(action = ArgAction::Count)]` with a field type of `u8`
   --> src/main.rs:139:43
    |
139 |     #[clap(short, long = "verbose", parse(from_occurrences))]
    |                                           ^^^^^^^^^^^^^^^^

+++ src/main.rs
@@ -14,7 +14,7 @@ use std::process;
 use std::str::FromStr;
 
 use chrono::{Date, Duration, NaiveDate, Utc};
-use clap::{ArgEnum, Parser, PossibleValue};
+use clap::{ArgAction, ArgEnum, Parser, PossibleValue};
 use colored::Colorize;
 use anyhow::{bail, Context};
 use log::debug;
@@ -136,8 +136,8 @@ struct Opts {
     )]
     timeout: Option<usize>,
 
-    #[clap(short, long = "verbose", parse(from_occurrences))]
-    verbosity: usize,
+    #[clap(short, long = "verbose", action = ArgAction::Count)]
+    verbosity: u8,
 
     #[clap(
         help = "Arguments to pass to cargo or the file specified by --script during tests",
